sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "starfire-plugin"
version = "0.1.0"
description = "SDK for building Starfire plugins (rules, skills, per-runtime adaptors)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Starfire" }]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/hongmingw/Starfire-AgentTeam"
Repository = "https://github.com/hongmingw/Starfire-AgentTeam"

[tool.setuptools.packages.find]
include = ["starfire_plugin*"]
exclude = ["template*", "tests*"]
